Implement a program to find the number of rabbits and chickens in a farm.
Problem Statement
Implement a program to find the number of rabbits and chickens in a farm. Given the number of heads and legs of the chickens and rabbits in a farm, identify and display the number of chickens and rabbits in the farm.

If the given input cannot make a valid number of rabbits and chickens, then display an appropriate message.