Implement a program to check whether a given number is a lucky number.
Problem Statement
Implement a program to check whether a given number is a lucky number.

A lucky number is a number whose sum of squares of every even-positioned digit (starting from the second position) is a multiple of 9.

E.g. - 1623 = 62+32 = 45 is a multiple of 9 and hence is a lucky number.